locality/ad-hoc-scripts/requirements.txt
requests
pillow
cairosvg